move all canvas things to renderer because its it responsability
