+++ src/MainLoop.js
@@ -6,7 +6,7 @@ import Keyboard from './Keyboard';
 const FRAMERATE = 60;
 const rAF = requestAnimationFrame || function(cb) { setTimeout(cb, 1 / FRAMERATE * 1000) };
 
-export default (level, canvasMovable, canvasStatic) => {
+export default (level) => {
 
     let lastTime;
     const loop = () => {
@@ -20,7 +20,7 @@ export default (level, canvasMovable, canvasStatic) => {
         // TODO use delta time in update
         Physics.update(level, deltaTime);
         Collision.check(level);
-        Renderer.render(level, canvasMovable, canvasStatic);
+        Renderer.render(level);
 
         //Check la fin du niveau (mort ou a atteint la porte)
         //c'est une collision dans tous les cas, a voir si on la fait
@@ -30,7 +30,7 @@ export default (level, canvasMovable, canvasStatic) => {
         rAF(loop);
     };
 
-    Renderer.init(level, canvasMovable, canvasStatic);
+    Renderer.init(level);
     loop();
 
 }
